Fix new linting errors
