+++ lib/assert.js
@@ -8,7 +8,6 @@ let errorProps = [
   "number",
   "stack"
 ];
-
 /**
   @hide
 */
